chore: rename di to runtime-context
