+++ packages/core/src/tools/types.ts
@@ -2,8 +2,8 @@ import type { ToolExecutionOptions, Tool } from 'ai';
 import type { ZodSchema, z } from 'zod';
 
 import type { IAction, IExecutionContext, MastraUnion } from '../action';
-import type { RuntimeContext } from '../di';
 import type { Mastra } from '../mastra';
+import type { RuntimeContext } from '../runtime-context';
 
 export type VercelTool = Tool;
