+++ packages/server/src/server/handlers/network.ts
@@ -1,4 +1,4 @@
-import type { RuntimeContext } from '@mastra/core/di';
+import type { RuntimeContext } from '@mastra/core/runtime-context';
 import type { AgentNetwork } from '@mastra/core/network';
 import { HTTPException } from '../http-exception';
 import type { Context } from '../types';
