+++ packages/server/src/server/handlers/tools.test.ts
@@ -1,5 +1,5 @@
 import { Agent } from '@mastra/core/agent';
-import { RuntimeContext } from '@mastra/core/di';
+import { RuntimeContext } from '@mastra/core/runtime-context';
 import { Mastra } from '@mastra/core/mastra';
 import { createTool } from '@mastra/core/tools';
 import type { ToolAction, VercelTool } from '@mastra/core/tools';
